requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
chromadb==0.5.5
duckdb==1.0.0
pyjwt==2.9.0
pydantic==2.9.0
rank-bm25==0.2.2
pyyaml==6.0.2
pandas==2.2.2